+++ src/main/java/com/itacademy/SpaceApplication.java
@@ -1,4 +1,4 @@
-package com.itacademy.game;
+package com.itacademy;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
