+++ src/main/java/com/itacademy/controller/UsersController.java
@@ -1,7 +1,8 @@
-package com.itacademy.game.controller.user;
+package com.itacademy.controller;
 
-import com.itacademy.game.entity.UserEntity;
-import com.itacademy.game.service.impl.UsersServiceImpl;
+import com.itacademy.entity.UserEntity;
+import com.itacademy.model.UserAuthModel;
+import com.itacademy.service.impl.UsersServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -14,15 +15,20 @@ public class UsersController {
     @Autowired
     private UsersServiceImpl usersService;
 
-    @GetMapping
+    @GetMapping("/getAll")
     public List<UserEntity> userMenu(Model model){
         return usersService.getAllUsers();
     }
 
-    @PostMapping
+    @PostMapping("/registration")
     public UserEntity newUser(@RequestBody UserEntity user){
         user= usersService.newUser(user);
         return user;
+    }
+
+    @PostMapping("/log-in")
+    public UserEntity getUser(@RequestBody UserAuthModel userAuthModel){
+        return usersService.getAuthorizet(userAuthModel);
     }
 
 }
