--- src/main/java/com/itacademy/game/service/UsersService.java
@@ -1,11 +1,0 @@
-package com.itacademy.game.service;
-
-import com.itacademy.game.entity.UserEntity;
-
-import java.util.List;
-
-public interface UsersService {
-    UserEntity newUser(UserEntity user);
-    List<UserEntity> getAllUsers();
-//    Users getByUserLogin(String login);
-}
