+++ src/test/java/com/itacademy/SpaceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.itacademy.game;
+package com.itacademy;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
